ALMOST Completed the backbone of the program
Here I have almost finished the backbone of the program with all the user options and all the functions. I need to complete the final visualisation of the platform comparison. Even though the original plan was nowhere near being perfectly executed, the change to the original idea made it so much easier.