Completed the whole program
Today I completed the whole program and everything works. There were some issues with plotting the data on matplotlib such as the bar graph not working when using the code from the sample program. This was fixed through a copy and spin off of the matplotlib website and now everything works.